Add files via upload
Purpose
The programmable Web and cloud is built on HTTP. In this lab we will utilize HTTP
programming to “GET” and crawl parts of the HTML based Web. The goal is to familiarize
students with HTTP and programmatically using it to access the web. Of course, the majority
of the access to the Cloud is done using APIs and development environments which target
specific services, these are built on HTTP.

Problem Statement
Create an application which takes two arguments:
1) A URL as a starting point
2) The number of hops from that URL (NumHops)